Add SWR test helpers
